jetstream::iroh::iroh::discovery

Trait Discovery

pub trait Discovery:
    Debug
    + Send
    + Sync
    + 'static {
    // Provided methods
    fn publish(&self, _data: &NodeData) { ... }
    fn resolve(
        &self,
        _node_id: PublicKey,
    ) -> Option<Pin<Box<dyn Stream<Item = Result<DiscoveryItem, DiscoveryError>> + Send>>> { ... }
}

Node discovery for super::Endpoint.

This trait defines publishing and resolving addressing information for a NodeId. This enables connecting to other nodes with only knowing the NodeId, by using this Discovery system to look up the actual addressing information. It is common for implementations to require each node to publish their own information before it can be looked up by other nodes.

The published addressing information can include both a RelayUrl and/or direct addresses. See NodeData for details.

To allow for discovery, the super::Endpoint will call publish whenever discovery information changes. If a discovery mechanism requires a periodic refresh, it should start its own task.

Provided Methods

fn publish(&self, _data: &NodeData)

Publishes the given NodeData to the discovery mechanism.

This is fire and forget, since the Endpoint can not wait for successful publishing. If publishing is async, the implementation should start it’s own task.

This will be called from a tokio task, so it is safe to spawn new tasks. These tasks will be run on the runtime of the super::Endpoint.

fn resolve( &self, _node_id: PublicKey, ) -> Option<Pin<Box<dyn Stream<Item = Result<DiscoveryItem, DiscoveryError>> + Send>>>

Resolves the DiscoveryItem for the given NodeId.

Once the returned [BoxStream] is dropped, the service should stop any pending work.