jetstream::iroh::iroh

Module dns

Configurable DNS resolver.

The main export is the DnsResolver struct. It provides methods to resolve domain names to IPv4 and IPv6 addresses, and to look up TXT records. Additionally, the resolver features methods to resolve the EndpointInfo for an iroh EndpointId from _iroh TXT records. See the crate::endpoint_info module documentation for details on how iroh endpoint records are structured.

Structs

Builder

Builder for DnsResolver.

DnsResolver

The DNS resolver used throughout iroh.

StaggeredError

Error returned when a staggered call fails.

TxtRecordData

Record data for a TXT record.

Enums

DnsError

Potential errors related to DNS operations.

DnsProtocol

Protocols over which DNS records can be resolved.

LookupError

Potential errors related to DNS endpoint address lookups.

Constants

DNS_TIMEOUT

Default DNS query timeout.

N0_DNS_ENDPOINT_ORIGIN_PROD

The n0 address lookup DNS origin, for production.

N0_DNS_ENDPOINT_ORIGIN_STAGING

The n0 address lookup DNS origin, for testing.

Traits

Resolver

Trait for DNS resolvers used in iroh.

Type Aliases

BoxIter

Boxed iterator alias.