jetstream::iroh::iroh

Module endpoint

The Endpoint allows establishing connections to other iroh endpoints.

The Endpoint is the main API interface to manage a local iroh endpoint. It allows connecting to and accepting connections from other endpoints. See the module docs for more details on how iroh connections work.

The main items in this module are:

Endpoint to establish iroh connections with other endpoints.

Builder to create an Endpoint.

Modules

presets

Presets allow configuring an endpoint quickly with a chosen set of defaults.

Structs

Accept

Future produced by Endpoint::accept.

AcceptBi

Future produced by [Connection::accept_bi]

AcceptUni

Future produced by [Connection::accept_uni]

Accepting

In-progress connection attempt future

AckFrequencyConfig

Parameters for controlling the peer’s acknowledgement frequency

ApplicationClose

Reason given by an application for closing the connection

BindOpts

Options when configuring binding an IP socket.

Builder

Builder for Endpoint.

Chunk

A chunk of data from the receive stream

ClosedStream

Error indicating that a stream has not been opened or has already been finished or reset

ConnectOptions

Options for the Endpoint::connect_with_opts function.

Connecting

An outgoing connection in progress.

Connection

A QUIC connection.

ConnectionClose

Reason given by the transport for closing the connection

ConnectionInfo

Information about a connection.

ConnectionStats

Connection statistics

ControllerMetrics

Common congestion controller metrics

CryptoError

Generic crypto errors

DirectAddr

A direct address on which an iroh-endpoint might be contactable.

Endpoint

Controls an iroh endpoint, establishing connections with other endpoints.

ExportKeyingMaterialError

Error returned by [Session::export_keying_material].

FrameStats

Number of frames transmitted or received of each frame type

HandshakeCompleted

Marker type for a connection that has completed the handshake.

IdleTimeout

Maximum duration of inactivity to accept before timing out the connection

Incoming

An incoming connection for which the server has not yet begun its parts of the handshake.

IncomingZeroRtt

Marker type for a connection that is in the incoming 0-RTT state.

Keys

A complete set of keys for a certain packet space

MtuDiscoveryConfig

Parameters governing MTU discovery.

OpenBi

Future produced by [Connection::open_bi]

OpenUni

Future produced by [Connection::open_uni]

OutgoingZeroRtt

Marker type for a connection that is in the outgoing 0-RTT state.

PathId

Id representing different paths when using multipath extension

PathInfo

Information about a network path used by a Connection.

PathInfoList

List of PathInfo for the network paths of a Connection.

PathStats

Statistics related to a transmission path

QuicTransportConfig

Parameters governing the core QUIC state machine

QuicTransportConfigBuilder

Builder for a QuicTransportConfig.

ReadDatagram

Future produced by [Connection::read_datagram]

RecvStream

A stream that can only be used to receive data

RemoteEndpointIdError

RemoteInfo

Information about a remote endpoint.

RetryError

Error for attempting to retry an Incoming which already bears a token from a previous retry

RttEstimator

RTT estimation for a particular network path

SendDatagram

Future produced by [Connection::send_datagram_wait]

SendStream

A stream that can only be used to send data

ServerConfig

Parameters governing incoming connections

ServerConfigBuilder

A builder for a ServerConfig.

StreamId

Identifier for a stream within a particular connection

TokenReuseError

Error for when a validation token may have been reused

TransportAddrInfo

Address of a remote with some metadata

TransportError

Transport-level errors occur when a peer violates the protocol specification

TransportErrorCode

Transport-level error code

TransportParameters

Transport parameters used to negotiate connection-level preferences between peers

UdpStats

Statistics about UDP datagrams transmitted or received on a connection

UnorderedRecvStream

A stream that can be used to receive data out-of-order.

UnsupportedVersion

Error indicating that the specified QUIC version is not supported

ValidationTokenConfig

Configuration for sending and handling validation tokens in incoming connections

VarInt

An integer less than 2^62

VarIntBoundsExceeded

Error returned when constructing a VarInt from a value >= 2^62

Written

Indicates how many bytes and chunks had been transferred in a write operation

Enums

AfterHandshakeOutcome

Outcome of EndpointHooks::after_handshake

AlpnError

AuthenticationError

BeforeConnectOutcome

Outcome of EndpointHooks::before_connect

BindError

ConnectError

ConnectWithOptsError

ConnectingError

ConnectionError

Reasons why a connection might be lost

Dir

Whether a stream communicates data in both directions or only from the initiator

DirectAddrType

The type of direct address.

FrameType

InvalidSocketAddr

QuicConnectError

Errors in the parameters being used to create a new connection

ReadError

Errors that arise from reading from a stream.

ReadExactError

Errors that arise from reading from a stream.

ReadToEndError

Errors from RecvStream::read_to_end

RelayMode

Configuration of the relay servers for an Endpoint.

ResetError

Errors that arise while waiting for a stream to be reset

SendDatagramError

Errors that can arise when sending a datagram

Side

Whether an endpoint was the initiator of a connection

Source

The origin or source through which an address associated with a remote endpoint was discovered.

StoppedError

Errors that arise while monitoring for a send stream stop from the peer

TransportAddrUsage

Information how a transport address is used.

WriteError

Errors that arise from writing to a stream

ZeroRttStatus

Returned from OutgoingZeroRttConnection::handshake_completed.

Constants

ENV_FORCE_STAGING_RELAYS

Environment variable to force the use of staging relays.

Traits

AeadKey

A key for sealing data with AEAD-based algorithms

ConnectionState

Trait to track the state of a Connection at compile time.

Controller

Common interface for different congestion controllers

ControllerFactory

Constructs controllers on demand

EndpointHooks

EndpointHooks intercept the connection establishment process of an Endpoint.

HandshakeTokenKey

A pseudo random key for HKDF

HeaderKey

Keys used to protect packet headers

PacketKey

Keys used to protect packet payloads

TimeSource

Object to get current SystemTime

ToSocketAddr

A simpler version of ToSocketAddrs, that does not do any DNS resolution.

TokenLog

Responsible for limiting clients’ ability to reuse validation tokens

Functions

default_relay_mode

Returns the default relay mode.

force_staging_infra

Returns true if the use of staging relays is forced.

Type Aliases

IncomingZeroRttConnection

A QUIC connection on the server-side that can possibly accept 0-RTT data.

OutgoingZeroRttConnection

The client side of a 0-RTT connection.