jetstream::iroh::iroh::endpoint

Struct QuicTransportConfigBuilder

pub struct QuicTransportConfigBuilder(/* private fields */);

Builder for a QuicTransportConfig.

Implementations

impl QuicTransportConfigBuilder

pub fn build(self) -> QuicTransportConfig

Build a QuicTransportConfig from the builder.

pub fn max_concurrent_bidi_streams( self, value: VarInt, ) -> QuicTransportConfigBuilder

Maximum number of incoming bidirectional streams that may be open concurrently.

Must be nonzero for the peer to open any bidirectional streams.

Worst-case memory use is directly proportional to max_concurrent_bidi_streams * stream_receive_window, with an upper bound proportional to receive_window.

pub fn max_concurrent_uni_streams( self, value: VarInt, ) -> QuicTransportConfigBuilder

Variant of max_concurrent_bidi_streams affecting unidirectional streams.

pub fn max_idle_timeout( self, value: Option<IdleTimeout>, ) -> QuicTransportConfigBuilder

Maximum duration of inactivity to accept before timing out the connection.

The true idle timeout is the minimum of this and the peer’s own max idle timeout. None represents an infinite timeout. Defaults to 30 seconds.

WARNING: If a peer or its network path malfunctions or acts maliciously, an infinite idle timeout can result in permanently hung futures!

let mut builder = QuicTransportConfig::builder()
    // Set the idle timeout as `VarInt`-encoded milliseconds
    .max_idle_timeout(Some(VarInt::from_u32(10_000).into()));

// Set the idle timeout as a `Duration`
builder = builder.max_idle_timeout(Some(Duration::from_secs(10).try_into()?));

let _cfg = builder.build();

pub fn stream_receive_window(self, value: VarInt) -> QuicTransportConfigBuilder

Maximum number of bytes the peer may transmit without acknowledgement on any one stream before becoming blocked.

This should be set to at least the expected connection latency multiplied by the maximum desired throughput. Setting this smaller than receive_window helps ensure that a single stream doesn’t monopolize receive buffers, which may otherwise occur if the application chooses not to read from a large stream for a time while still requiring data on other streams.

pub fn receive_window(self, value: VarInt) -> QuicTransportConfigBuilder

Maximum number of bytes the peer may transmit across all streams of a connection before becoming blocked.

This should be set to at least the expected connection latency multiplied by the maximum desired throughput. Larger values can be useful to allow maximum throughput within a stream while another is blocked.

pub fn send_window(self, value: u64) -> QuicTransportConfigBuilder

Maximum number of bytes to transmit to a peer without acknowledgment.

Provides an upper bound on memory when communicating with peers that issue large amounts of flow control credit. Endpoints that wish to handle large numbers of connections robustly should take care to set this low enough to guarantee memory exhaustion does not occur if every connection uses the entire window.

pub fn send_fairness(self, value: bool) -> QuicTransportConfigBuilder

Whether to implement fair queuing for send streams having the same priority.

When enabled, connections schedule data from outgoing streams having the same priority in a round-robin fashion. When disabled, streams are scheduled in the order they are written to.

Note that this only affects streams with the same priority. Higher priority streams always take precedence over lower priority streams.

Disabling fairness can reduce fragmentation and protocol overhead for workloads that use many small streams.

pub fn packet_threshold(self, value: u32) -> QuicTransportConfigBuilder

Maximum reordering in packet number space before FACK style loss detection considers a packet lost. Should not be less than 3, per RFC5681.

pub fn time_threshold(self, value: f32) -> QuicTransportConfigBuilder

Maximum reordering in time space before time based loss detection considers a packet lost, as a factor of RTT.

pub fn initial_rtt(self, value: Duration) -> QuicTransportConfigBuilder

The RTT used before an RTT sample is taken.

pub fn initial_mtu(self, value: u16) -> QuicTransportConfigBuilder

The initial value to be used as the maximum UDP payload size before running MTU discovery (see QuicTransportConfigBuilder::mtu_discovery_config).

Must be at least 1200, which is the default, and known to be safe for typical internet applications. Larger values are more efficient, but increase the risk of packet loss due to exceeding the network path’s IP MTU. If the provided value is higher than what the network path actually supports, packet loss will eventually trigger black hole detection and bring it down to QuicTransportConfigBuilder::min_mtu.

pub fn min_mtu(self, value: u16) -> QuicTransportConfigBuilder

The maximum UDP payload size guaranteed to be supported by the network.

Must be at least 1200, which is the default, and lower than or equal to QuicTransportConfigBuilder::initial_mtu.

Real-world MTUs can vary according to ISP, VPN, and properties of intermediate network links outside of either endpoint’s control. Extreme care should be used when raising this value outside of private networks where these factors are fully controlled. If the provided value is higher than what the network path actually supports, the result will be unpredictable and catastrophic packet loss, without a possibility of repair. Prefer QuicTransportConfigBuilder::initial_mtu together with QuicTransportConfigBuilder::mtu_discovery_config to set a maximum UDP payload size that robustly adapts to the network.

pub fn mtu_discovery_config( self, value: Option<MtuDiscoveryConfig>, ) -> QuicTransportConfigBuilder

Specifies the MTU discovery config (see MtuDiscoveryConfig for details).

Enabled by default.

pub fn pad_to_mtu(self, value: bool) -> QuicTransportConfigBuilder

Pad UDP datagrams carrying application data to current maximum UDP payload size.

Disabled by default. UDP datagrams containing loss probes are exempt from padding.

Enabling this helps mitigate traffic analysis by network observers, but it increases bandwidth usage. Without this mitigation precise plain text size of application datagrams as well as the total size of stream write bursts can be inferred by observers under certain conditions. This analysis requires either an uncongested connection or application datagrams too large to be coalesced.

pub fn ack_frequency_config( self, value: Option<AckFrequencyConfig>, ) -> QuicTransportConfigBuilder

Specifies the ACK frequency config (see AckFrequencyConfig for details).

The provided configuration will be ignored if the peer does not support the acknowledgement frequency QUIC extension.

Defaults to None, which disables controlling the peer’s acknowledgement frequency. Even if set to None, the local side still supports the acknowledgement frequency QUIC extension and may use it in other ways.

pub fn persistent_congestion_threshold( self, value: u32, ) -> QuicTransportConfigBuilder

Number of consecutive PTOs after which network is considered to be experiencing persistent congestion.

pub fn keep_alive_interval(self, value: Duration) -> QuicTransportConfigBuilder

Period of inactivity before sending a keep-alive packet.

Keep-alive packets prevent an inactive but otherwise healthy connection from timing out.

None to disable, which is the default. Only one side of any given connection needs keep-alive enabled for the connection to be preserved. Must be set lower than the idle_timeout of both peers to be effective.

pub fn crypto_buffer_size(self, value: usize) -> QuicTransportConfigBuilder

Maximum quantity of out-of-order crypto layer data to buffer.

pub fn allow_spin(self, value: bool) -> QuicTransportConfigBuilder

Whether the implementation is permitted to set the spin bit on this connection.

This allows passive observers to easily judge the round trip time of a connection, which can be useful for network administration but sacrifices a small amount of privacy.

pub fn datagram_receive_buffer_size( self, value: Option<usize>, ) -> QuicTransportConfigBuilder

Maximum number of incoming application datagram bytes to buffer, or None to disable incoming datagrams.

The peer is forbidden to send single datagrams larger than this size. If the aggregate size of all datagrams that have been received from the peer but not consumed by the application exceeds this value, old datagrams are dropped until it is no longer exceeded.

pub fn datagram_send_buffer_size( self, value: usize, ) -> QuicTransportConfigBuilder

Maximum number of outgoing application datagram bytes to buffer.

While datagrams are sent ASAP, it is possible for an application to generate data faster than the link, or even the underlying hardware, can transmit them. This limits the amount of memory that may be consumed in that case. When the send buffer is full and a new datagram is sent, older datagrams are dropped until sufficient space is available.

pub fn congestion_controller_factory( self, factory: Arc<dyn ControllerFactory + Send + Sync>, ) -> QuicTransportConfigBuilder

How to construct new congestion::Controllers.

Typically the refcounted configuration of a congestion::Controller, e.g. a congestion::NewRenoConfig.

Example

let config = QuicTransportConfig::builder()
    .congestion_controller_factory(Arc::new(congestion::NewRenoConfig::default()))
    .build();

pub fn enable_segmentation_offload( self, enabled: bool, ) -> QuicTransportConfigBuilder

Whether to use “Generic Segmentation Offload” to accelerate transmits, when supported by the environment.

Defaults to true.

GSO dramatically reduces CPU consumption when sending large numbers of packets with the same headers, such as when transmitting bulk data on a connection. However, it is not supported by all network interface drivers or packet inspection tools. quinn-udp will attempt to disable GSO automatically when unavailable, but this can lead to spurious packet loss at startup, temporarily degrading performance.

pub fn send_observed_address_reports( self, enabled: bool, ) -> QuicTransportConfigBuilder

Whether to send observed address reports to peers.

This will aid peers in inferring their reachable address, which in most NATd networks will not be easily available to them.

pub fn receive_observed_address_reports( self, enabled: bool, ) -> QuicTransportConfigBuilder

Whether to receive observed address reports from other peers.

Peers with the address discovery extension enabled that are willing to provide observed address reports will do so if this transport parameter is set. In general, observed address reports cannot be trusted. This, however, can aid the current endpoint in inferring its reachable address, which in most NATd networks will not be easily available.

pub fn max_concurrent_multipath_paths( self, max_concurrent: u32, ) -> QuicTransportConfigBuilder

Enables the Multipath Extension for QUIC.

Setting this to any nonzero value will enable the Multipath Extension for QUIC, https://datatracker.ietf.org/doc/draft-ietf-quic-multipath/.

The value provided specifies the number maximum number of paths this endpoint may open concurrently when multipath is negotiated. For any path to be opened, the remote must enable multipath as well.

Note: this method will ignore values less than the recommended 13 and will log a warning.

pub fn default_path_max_idle_timeout( self, timeout: Duration, ) -> QuicTransportConfigBuilder

Sets a default per-path maximum idle timeout.

If the path is idle for this long the path will be abandoned. Bear in mind this will interact with the QuicTransportConfigBuilder::max_idle_timeout, if the last path is abandoned the entire connection will be closed.

Note: this method will ignore values higher than the recommended 6500 ms and will log a warning.

pub fn default_path_keep_alive_interval( self, interval: Duration, ) -> QuicTransportConfigBuilder

Sets a default per-path keep alive interval.

Note that this does not interact with the connection-wide QuicTransportConfigBuilder::keep_alive_interval. This setting will keep this path active, QuicTransportConfigBuilder::keep_alive_interval will keep the connection active, with no control over which path is used for this.

Note: this method will ignore values higher than the recommended 5 seconds and will log a warning.

pub fn set_max_remote_nat_traversal_addresses( self, max_addresses: u8, ) -> QuicTransportConfigBuilder

Sets the maximum number of nat traversal addresses this endpoint allows the remote to advertise.

Setting this to any nonzero value will enable Iroh’s holepunching, loosely based in the Nat Traversal Extension for QUIC, see https://www.ietf.org/archive/id/draft-seemann-quic-nat-traversal-02.html

This implementation expects the multipath extension to be enabled as well. If not yet enabled via Self::max_concurrent_multipath_paths, a default value of 12 will be used.

Note: this method will ignore values less than the recommended 12 and will log a warning.