jetstream::iroh::iroh::endpoint

Type Alias OutgoingZeroRttConnection

pub type OutgoingZeroRttConnection = Connection<OutgoingZeroRtt>;

The client side of a 0-RTT connection.

This is created using Connecting::into_0rtt.

Creating a OutgoingZeroRttConnection means that the endpoint is capable of attempting a 0-RTT connection with the remote. The remote may still reject the 0-RTT connection. In which case, any data sent before the handshake has completed may need to be resent.

Look at the OutgoingZeroRttConnection::handshake_completed method for more details.

Aliased Type

pub struct OutgoingZeroRttConnection { /* private fields */ }