jetstream::iroh::iroh::protocol

Trait ProtocolHandler

pub trait ProtocolHandler:
    Send
    + Sync
    + Debug
    + 'static {
    // Required method
    fn accept(
        &self,
        connection: Connection,
    ) -> impl Future<Output = Result<(), AcceptError>> + Send;

    // Provided methods
    fn on_connecting(
        &self,
        connecting: Connecting,
    ) -> impl Future<Output = Result<Connection, AcceptError>> + Send { ... }
    fn shutdown(&self) -> impl Future<Output = ()> + Send { ... }
}

Handler for incoming connections.

A router accepts connections for arbitrary ALPN protocols.

With this trait, you can handle incoming connections for any protocol.

Implement this trait on a struct that should handle incoming connections. The protocol handler must then be registered on the node for an ALPN protocol with crate::protocol::RouterBuilder::accept.

See the module documentation for an example.

Required Methods

fn accept( &self, connection: Connection, ) -> impl Future<Output = Result<(), AcceptError>> + Send

Handle an incoming connection.

Can be implemented as async fn accept(&self, connection: Connection) -> Result<Connection>.

The returned future runs on a freshly spawned tokio task so it can be long-running.

When Router::shutdown is called, no further connections will be accepted, and the futures returned by Self::accept will be aborted after the future returned from ProtocolHandler::shutdown completes.

Provided Methods

fn on_connecting( &self, connecting: Connecting, ) -> impl Future<Output = Result<Connection, AcceptError>> + Send

Optional interception point to handle the Connecting state.

Can be implemented as async fn on_connecting(&self, connecting: Connecting) -> Result<Connection>.

This enables accepting 0-RTT data from clients, among other things.

fn shutdown(&self) -> impl Future<Output = ()> + Send

Called when the router shuts down.

Can be implemented as async fn shutdown(&self).

This is called from Router::shutdown. The returned future is awaited before the router closes the endpoint.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.