jetstream

Module prelude

Re-exports

pub extern crate async_trait;

pub extern crate futures;

pub extern crate jetstream_macros;

pub extern crate jetstream_rpc;

pub extern crate jetstream_wireformat;

pub extern crate lazy_static;

pub extern crate trait_variant;

Modules

client

server

Macros

lazy_static

Structs

Data

A type that encodes an arbitrary number of bytes of data. Typically used for Rread Twrite messages. This differs from a Vec<u8> in that it encodes the number of bytes using a u32 instead of a u16.

Frame

Framed

A unified Stream and Sink interface to an underlying I/O object, using the Encoder and Decoder traits to encode and decode frames.

Tag

Enums

Error

Traits

ClientTransport

Framer

LazyStatic

Support trait for enabling a few common operation on lazy static values.

Message

A trait representing a message that can be encoded and decoded.

Protocol

Defines the request and response types for the JetStream protocol.

WireFormat

A type that can be encoded on the wire using the 9P protocol.

__Deref

Used for immutable dereferencing operations, like *v.

Functions

initialize

Takes a shared reference to a lazy static and initializes it if it has not been already.

Attribute Macros

async_trait

make

Creates a specialized version of a base trait that adds bounds to async fn and/or -> impl Trait return types.

service

Service attribute macro for creating RPC services

Derive Macros

JetStreamWireFormat

Derives wire format encoding for structs