jetstream::prelude::callsite

Trait Callsite

pub trait Callsite: Sync {
    // Required methods
    fn set_interest(&self, interest: Interest);
    fn metadata(&self) -> &Metadata<'_>;
}

Trait implemented by callsites.

These functions are only intended to be called by the callsite registry, which correctly handles determining the common interest between all subscribers.

See the module-level documentation for details on callsites.

Required Methods

fn set_interest(&self, interest: Interest)

Sets the Interest for this callsite.

See the documentation on callsite interest caching for details.

fn metadata(&self) -> &Metadata<'_>

Returns the metadata associated with the callsite.

Note: Implementations of this method should not produce Metadata
that share the same callsite Identifier but otherwise differ in any
way (e.g., have different names).