jetstream::websocket::tokio_tungstenite

Function client_async_tls_with_config

pub async fn client_async_tls_with_config<R, S>(
    request: R,
    stream: S,
    config: Option<WebSocketConfig>,
    connector: Option<Connector>,
) -> Result<(WebSocketStream<MaybeTlsStream<S>>, Response<Option<Vec<u8>>>), Error>
where
    R: IntoClientRequest + Unpin,
    S: 'static + AsyncRead + AsyncWrite + Send + Unpin,
    MaybeTlsStream<S>: Unpin,

The same as client_async_tls() but the one can specify a websocket configuration, and an optional connector. If no connector is specified, a default one will be created.

Please refer to client_async_tls() for more details.