jetstream::websocket::tokio_tungstenite::tungstenite::client

Function client

pub fn client<Stream, Req>(
    request: Req,
    stream: Stream,
) -> Result<(WebSocket<Stream>, Response<Option<Vec<u8>>>), HandshakeError<ClientHandshake<Stream>>>
where
    Stream: Read + Write,
    Req: IntoClientRequest,

Do the client handshake over the given stream.

Use this function if you need a nonblocking handshake support or if you want to use a custom stream like mio::net::TcpStream or openssl::ssl::SslStream. Any stream supporting Read + Write will do.