jetstream::websocket::tokio_tungstenite::tungstenite::protocol

Struct WebSocketConfig

#[non_exhaustive]
pub struct WebSocketConfig {
    pub read_buffer_size: usize,
    pub write_buffer_size: usize,
    pub max_write_buffer_size: usize,
    pub max_message_size: Option<usize>,
    pub max_frame_size: Option<usize>,
    pub accept_unmasked_frames: bool,
}

The configuration for WebSocket connection.

Example

let conf = WebSocketConfig::default()
    .read_buffer_size(256 * 1024)
    .write_buffer_size(256 * 1024);

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

read_buffer_size: usize

Read buffer capacity. This buffer is eagerly allocated and used for receiving messages.

For high read load scenarios a larger buffer, e.g. 128 KiB, improves performance.

For scenarios where you expect a lot of connections and donโt need high read load performance a smaller buffer, e.g. 4 KiB, would be appropriate to lower total memory usage.

The default value is 128 KiB.

write_buffer_size: usize

The target minimum size of the write buffer to reach before writing the data to the underlying stream. The default value is 128 KiB.

If set to 0 each message will be eagerly written to the underlying stream. It is often more optimal to allow them to buffer a little, hence the default value.

Note: flush will always fully write the buffer regardless.

max_write_buffer_size: usize

The max size of the write buffer in bytes. Setting this can provide backpressure in the case the write buffer is filling up due to write errors. The default value is unlimited.

Note: The write buffer only builds up past write_buffer_size when writes to the underlying stream are failing. So the write buffer can not fill up if you are not observing write errors even if not flushing.

Note: Should always be at least write_buffer_size + 1 message and probably a little more depending on error handling strategy.

max_message_size: Option<usize>

The maximum size of an incoming message. None means no size limit. The default value is 64 MiB which should be reasonably big for all normal use-cases but small enough to prevent memory eating by a malicious user.

max_frame_size: Option<usize>

The maximum size of a single incoming message frame. None means no size limit. The limit is for frame payload NOT including the frame header. The default value is 16 MiB which should be reasonably big for all normal use-cases but small enough to prevent memory eating by a malicious user.

accept_unmasked_frames: bool

When set to true, the server will accept and handle unmasked frames from the client. According to the RFC 6455, the server must close the connection to the client in such cases, however it seems like there are some popular libraries that are sending unmasked frames, ignoring the RFC. By default this option is set to false, i.e. according to RFC 6455.

Implementations

impl WebSocketConfig

pub fn read_buffer_size(self, read_buffer_size: usize) -> WebSocketConfig

Set Self::read_buffer_size.

pub fn write_buffer_size(self, write_buffer_size: usize) -> WebSocketConfig

Set Self::write_buffer_size.

pub fn max_write_buffer_size( self, max_write_buffer_size: usize, ) -> WebSocketConfig

Set Self::max_write_buffer_size.

pub fn max_message_size( self, max_message_size: Option<usize>, ) -> WebSocketConfig

Set Self::max_message_size.

pub fn max_frame_size(self, max_frame_size: Option<usize>) -> WebSocketConfig

Set Self::max_frame_size.

pub fn accept_unmasked_frames( self, accept_unmasked_frames: bool, ) -> WebSocketConfig

Set Self::accept_unmasked_frames.