Crate jetstream

JetStream

JetStream, is a collection of crates that provide a set of tools to build distributed systems. It started it’s life off in the CrosVM project, and has since been extracted into it’s own project. For more information please see the JetStream Book

Modules

http

iroh

macros

p9

prelude

quic

Macros

prost_wireformat

Creates a newtype wrapper around a prost Message type that implements WireFormat.