jetstream::iroh::iroh::endpoint

Struct BindOpts

pub struct BindOpts { /* private fields */ }

Options when configuring binding an IP socket.

Implementations

impl BindOpts

pub fn set_prefix_len(self, prefix_len: u8) -> BindOpts

Sets the network prefix length of the subnet this interface is in.

The subnets of bound sockets are used to route outgoing datagrams not belonging to an existing traffic flow to the socket they should be sent on. Subnets are ordered from longest prefix length to shortest prefix length and the first subnet which contains the destination IP address will be chosen. If no subnet matches but there is a bound socket marked with Self::set_is_default_route then this socket will be used. In this case it is assumed the attached subnet has a gateway router to forward the datagram.

Defaults to 0, which means all IP addresses will belong to the subnet of this socketโs address. If multiple sockets of the same address family (IPv4 or IPv6) are bound with such a /0 prefix the socket which will be chosen is undefined.

For IPv4 sockets the maximum prefix is 32. For IPv6 the maximum prefix is 128.

pub fn prefix_len(&self) -> u8

Returns the prefix_len, see Self::set_prefix_len.

pub fn set_is_required(self, is_required: bool) -> BindOpts

Sets whether bind errors are fatal for this socket.

If false and this socket fails to bind, the error will be silently ignored and the endpoint will still be created.

Defaults to true.

pub fn is_required(&self) -> bool

Returns the value set by Self::set_is_required.

pub fn set_is_default_route(self, is_default_route: bool) -> BindOpts

Sets whether this is a default route.

The default route is used for outgoing datagrams not belonging to an existing traffic flow, which does not fit in any subnet of the bound sockets. It is assumed this subnet has a gateway router to route such packets.

See Self::set_prefix_len for details on how this routing works.

If not set explicitly, then Self::is_default_route will return true if the prefix length is set to 0 (the default) and false otherwise.

pub fn is_default_route(&self) -> bool

Returns whether this is a default route.

If Self::set_is_default_route has been called then that value is returned. Otherwise, returns true if the prefix_len is 0 and false otherwise.