jetstream::iroh::iroh::endpoint

Struct Builder

pub struct Builder { /* private fields */ }

Builder for Endpoint.

By default the endpoint will generate a new random SecretKey, which will result in a new NodeId.

To create the Endpoint call Builder::bind.

Implementations

impl Builder

pub async fn bind(self) -> Result<Endpoint, BindError>

Binds the magic endpoint.

pub fn bind_addr_v4(self, addr: SocketAddrV4) -> Builder

Sets the IPv4 bind address.

Setting the port to 0 will use a random port. If the port specified is already in use, it will fallback to choosing a random port.

By default will use 0.0.0.0:0 to bind to.

pub fn bind_addr_v6(self, addr: SocketAddrV6) -> Builder

Sets the IPv6 bind address.

Setting the port to 0 will use a random port. If the port specified is already in use, it will fallback to choosing a random port.

By default will use [::]:0 to bind to.

pub fn secret_key(self, secret_key: SecretKey) -> Builder

Sets a secret key to authenticate with other peers.

This secret key’s public key will be the PublicKey of this endpoint and thus also its NodeId

If not set, a new secret key will be generated.

pub fn alpns(self, alpn_protocols: Vec<Vec<u8>>) -> Builder

Sets the ALPN protocols that this endpoint will accept on incoming connections.

Not setting this will still allow creating connections, but to accept incoming connections at least one ALPN must be set.

pub fn relay_mode(self, relay_mode: RelayMode) -> Builder

Sets the relay servers to assist in establishing connectivity.

Relay servers are used to establish initial connection with another iroh node. They also perform various functions related to hole punching, see the crate docs for more details.

By default the number 0 relay servers are used, see RelayMode::Default.

When using RelayMode::Custom, the provided relay_map must contain at least one configured relay node. If an invalid RelayMap is provided bind will result in an error.

pub fn clear_discovery(self) -> Builder

Removes all discovery services from the builder.

pub fn discovery(self, discovery: impl IntoDiscovery) -> Builder

Optionally sets a discovery mechanism for this endpoint.

If you want to combine multiple discovery services, you can use Builder::add_discovery instead. This will internally create a crate::discovery::ConcurrentDiscovery.

If no discovery service is set, connecting to a node without providing its direct addresses or relay URLs will fail.

See the documentation of the crate::discovery::Discovery trait for details.

pub fn add_discovery(self, discovery: impl IntoDiscovery) -> Builder

Adds a discovery mechanism for this endpoint.

The function discovery will be called on endpoint creation with the configured secret key of the endpoint. Discovery services that need to publish information need to use this secret key to sign the information.

If you add multiple discovery services, they will be combined using a crate::discovery::ConcurrentDiscovery.

If no discovery service is set, connecting to a node without providing its direct addresses or relay URLs will fail.

To clear all discovery services, use Builder::clear_discovery.

See the documentation of the crate::discovery::Discovery trait for details.

pub fn discovery_n0(self) -> Builder

Configures the endpoint to use the default n0 DNS discovery service.

The default discovery service publishes to and resolves from the n0.computer dns server iroh.link.

This is equivalent to adding both a crate::discovery::pkarr::PkarrPublisher and a crate::discovery::dns::DnsDiscovery, both configured to use the n0.computer dns server.

This will by default use N0_DNS_PKARR_RELAY_PROD. When in tests, or when the test-utils feature is enabled, this will use the N0_DNS_PKARR_RELAY_STAGING.

pub fn user_data_for_discovery(self, user_data: UserData) -> Builder

Sets the initial user-defined data to be published in discovery services for this node.

When using discovery services, this string of UserData will be published together with the node’s addresses and relay URL. When other nodes discover this node, they retrieve the UserData in addition to the addressing info.

Iroh itself does not interpret the user-defined data in any way, it is purely left for applications to parse and use.

pub fn transport_config(self, transport_config: TransportConfig) -> Builder

Sets a custom quinn::TransportConfig for this endpoint.

The transport config contains parameters governing the QUIC state machine.

If unset, the default config is used. Default values should be suitable for most internet applications. Applications protocols which forbid remotely-initiated streams should set max_concurrent_bidi_streams and max_concurrent_uni_streams to zero.

Please be aware that changing some settings may have adverse effects on establishing and maintaining direct connections.

pub fn dns_resolver(self, dns_resolver: DnsResolver) -> Builder

Optionally sets a custom DNS resolver to use for this endpoint.

The DNS resolver is used to resolve relay hostnames, and node addresses if crate::discovery::dns::DnsDiscovery is configured.

By default, a new DNS resolver is created which is configured to use the host system’s DNS configuration. You can pass a custom instance of DnsResolver here to use a differently configured DNS resolver for this endpoint, or to share a DnsResolver between multiple endpoints.

pub fn proxy_url(self, url: Url) -> Builder

Sets an explicit proxy url to proxy all HTTP(S) traffic through.

pub fn proxy_from_env(self) -> Builder

Sets the proxy url from the environment, in this order:

HTTP_PROXY

http_proxy

HTTPS_PROXY

https_proxy

pub fn keylog(self, keylog: bool) -> Builder

Enables saving the TLS pre-master key for connections.

This key should normally remain secret but can be useful to debug networking issues by decrypting captured traffic.

If keylog is true then setting the SSLKEYLOGFILE environment variable to a filename will result in this file being used to log the TLS pre-master keys.

pub fn max_tls_tickets(self, n: usize) -> Builder

Set the maximum number of TLS tickets to cache.

Set this to a larger value if you want to do 0rtt connections to a large number of clients.

The default is 256, taking about 150 KiB in memory.