jetstream

Module prelude

Re-exports

pub extern crate async_trait;

pub extern crate futures;

pub extern crate jetstream_macros;

pub extern crate jetstream_rpc;

pub extern crate jetstream_wireformat;

pub extern crate lazy_static;

pub extern crate tracing;

pub extern crate tracing_subscriber;

pub extern crate trait_variant;

Modules

callsite

Callsites represent the source locations from which spans or events originate.

client

dispatcher

Dispatches trace events to Subscribers.

event

Events represent single points in time during the execution of a program.

field

Span and Event key-value data.

instrument

Attach a span to a std::future::Future.

level_filters

Trace verbosity level filtering.

metadata

Metadata describing trace data.

server

span

Spans represent periods of time in which a program was executing in a particular context.

subscriber

Collects and records trace data.

Macros

debug

Constructs an event at the debug level.

debug_span

Constructs a span at the debug level.

enabled

Checks whether a span or event is enabled based on the provided metadata.

error

Constructs an event at the error level.

error_span

Constructs a span at the error level.

event

Constructs a new Event.

event_enabled

Tests whether an event with the specified level and target would be enabled.

info

Constructs an event at the info level.

info_span

Constructs a span at the info level.

lazy_static

metadata

Statically constructs new span metadata.

span

Constructs a new span.

span_enabled

Tests whether a span with the specified level and target would be enabled.

trace

Constructs an event at the trace level.

trace_span

Constructs a span at the trace level.

warn

Constructs an event at the warn level.

warn_span

Constructs a span at the warn level.

Structs

Context

Data

A type that encodes an arbitrary number of bytes of data. Typically used for Rread Twrite messages. This differs from a Vec<u8> in that it encodes the number of bytes using a u32 instead of a u16.

Dispatch

Dispatch trace data to a Subscriber.

Event

Events represent single points in time where something occurred during the execution of a program.

Frame

Framed

A unified Stream and Sink interface to an underlying I/O object, using the Encoder and Decoder traits to encode and decode frames.

Id

Identifies a span within the context of a subscriber.

Level

Describes the level of verbosity of a span or event.

Metadata

Metadata describing a span or event.

Span

A handle representing a span, with the capability to enter the span if it exists.

Tag

Enums

Error

Traits

Callsite

Trait implemented by callsites.

ClientTransport

Framer

Instrument

Attaches spans to a std::future::Future.

LazyStatic

Support trait for enabling a few common operation on lazy static values.

Message

A trait representing a message that can be encoded and decoded.

Protocol

Defines the request and response types for the JetStream protocol.

Subscriber

Trait representing the functions required to collect trace data.

Value

A field value of an erased type.

WireFormat

A type that can be encoded on the wire using the 9P protocol.

__Deref

Used for immutable dereferencing operations, like *v.

Functions

initialize

Takes a shared reference to a lazy static and initializes it if it has not been already.

Attribute Macros

async_trait

instrument

Instruments a function to create and enter a tracing span every time the function is called.

make

Creates a specialized version of a base trait that adds bounds to async fn and/or -> impl Trait return types.

service

Service attribute macro for creating RPC services

Derive Macros

JetStreamWireFormat

Derives wire format encoding for structs